\name{DescendMethods-package}
\alias{DescendMethods-package}
\alias{DescendMethods}
\docType{package}
\title{\packageTitle{DescendMethods}}
\description{\packageDescription{DescendMethods}}
\section{Package Content}{\packageIndices{DescendMethods}}
\author{\packageAuthor{DescendMethods}}
\section{Maintainer}{\packageMaintainer{DescendMethods}}
\keyword{package}
