% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientDescend.R
\name{gd}
\alias{gd}
\title{Gradient Descend}
\usage{
gd(data, stepsize = 1e-04, init = NULL, tol = 1e-04, maxit = 1000L, verb = F)
}
\arguments{
\item{data}{}

\item{stepsize}{}

\item{init}{}

\item{tol}{}

\item{maxit}{}

\item{verb}{}
}
\value{

}
\description{
Gradient Descend
}
