% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientDescend.R
\name{LossD}
\alias{LossD}
\title{LossD}
\usage{
LossD(b, X, Y)
}
\arguments{
\item{b}{\link{vector} the vector of \mjseqn{\beta} parameters in the formula above.}

\item{X}{\link{matrix} the design matrix \code{X} in the formula above.}

\item{Y}{\link{vector} the response vector \code{Y} in the formula above.}
}
\value{
\link{double} the function evaluation
}
\description{
\loadmathjax This function evaluates the following quadratic loss function
\mjsdeqn{L(\beta) = (X\beta - Y)^2}
}
