% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{PcrossVD}
\alias{PcrossVD}
\title{PcrossVD}
\usage{
PcrossVD(data, K = NULL, get_mean = T, n_clust = NULL, OPT, ...)
}
\arguments{
\item{data}{\link{list} containing the data, elements must be named \code{X} and \code{Y}.}

\item{n_clust}{\link{integer} the number of clusters to use. If \code{NULL}, the function selects all available cores but one.}
}
\value{
if \code{get_mean} is \code{TRUE}, the CV-MSE is returned.
Otherwise the function returns a \link{vector} containing all MSE computed for each fold.
}
\description{
This function is the parallel version of \code{\link{CrossVD}}.
}
