% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteepestDescend.R
\name{sdg}
\alias{sdg}
\title{Steepest Descend}
\usage{
sdg(data, init = NULL, tol = 1e-04, maxit = 1000L, verb = F)
}
\arguments{
\item{data}{}

\item{init}{}

\item{tol}{}

\item{maxit}{}

\item{verb}{}
}
\value{

}
\description{
Steepest Descend
}
