% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteepestDescend.R
\name{SteepD}
\alias{SteepD}
\title{Steepest Descend}
\usage{
SteepD(data, init = NULL, tol = 1e-04, maxit = 1000L, verb = F, check_loss = F)
}
\arguments{
\item{data}{}

\item{init}{}

\item{tol}{}

\item{maxit}{}

\item{verb}{}

\item{check_loss}{}
}
\value{

}
\description{
Steepest Descend
}
