% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CrossVD}
\alias{CrossVD}
\title{CrossVD}
\usage{
CrossVD(data, K = NULL, get_mean = T, OPT, ...)
}
\arguments{
\item{data}{}

\item{K}{}

\item{...}{}

\item{FUN}{}
}
\value{

}
\description{
CrossVD
}
