% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientDescend.R
\name{GradD}
\alias{GradD}
\title{Gradient Descend}
\usage{
GradD(
  data,
  stepsize = 1e-04,
  init = NULL,
  tol = 1e-04,
  maxit = 1000L,
  verb = F,
  check_loss = F
)
}
\arguments{
\item{data}{}

\item{stepsize}{}

\item{init}{}

\item{tol}{}

\item{maxit}{}

\item{verb}{}

\item{check_loss}{}
}
\value{

}
\description{
Gradient Descend
}
