% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{PcrossVD}
\alias{PcrossVD}
\title{PcrossVD}
\usage{
PcrossVD(data, K = NULL, get_mean = T, n_clust = NULL, OPT, ...)
}
\arguments{
\item{data}{}

\item{K}{}

\item{get_mean}{}

\item{OPT}{}

\item{...}{}
}
\value{

}
\description{
PcrossVD
}
